+++ src/test/java/com/deanofwalls/bachelor_strength/TestCaloriesForHypertrophy.java
@@ -18,10 +18,10 @@ public class TestCaloriesForHypertrophy {
     void testCaloriesForHypertrophy() {
         //Given: setup our test data
         UserModel userModel = new UserModel(30, 180, 170, "male");
-        double expectedCaloriesForHypertrophy = 2714 + 500; //rounded value
+        int expectedCaloriesForHypertrophy = 2714 + 500; //rounded value
 
         //When
-        double actualCaloriesForHypertrophy = fitnessService.calculateCaloriesForHypertrophy(userModel);
+        int actualCaloriesForHypertrophy = fitnessService.calculateCaloriesForHypertrophy(userModel);
 
         //Then
         assertEquals(expectedCaloriesForHypertrophy, actualCaloriesForHypertrophy, 0.01, "The CaloriesForHypertrophy should match the expected value.");
